+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/afericao/_model/Afericao.java
@@ -3,10 +3,11 @@ package br.com.zalf.prolog.webservice.frota.pneu.afericao._model;
 import br.com.zalf.prolog.webservice.commons.gson.Exclude;
 import br.com.zalf.prolog.webservice.commons.gson.RuntimeTypeAdapterFactory;
 import br.com.zalf.prolog.webservice.frota.pneu._model.Pneu;
+import br.com.zalf.prolog.webservice.frota.pneu.afericao.configuracao._model.FormaColetaDadosAfericaoEnum;
 import br.com.zalf.prolog.webservice.gente.colaborador.model.Colaborador;
 import br.com.zalf.prolog.webservice.geral.unidade._model.Unidade;
-import lombok.Data;
 import org.jetbrains.annotations.NotNull;
+import lombok.Data;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -34,6 +35,7 @@ public abstract class Afericao {
      */
     private long tempoRealizacaoAfericaoInMillis;
     private TipoMedicaoColetadaAfericao tipoMedicaoColetadaAfericao;
+    private FormaColetaDadosAfericaoEnum formaColetaDadosAfericao;
 
     @NotNull
     public static RuntimeTypeAdapterFactory<Afericao> provideTypeAdapterFactory() {